requirements.txt
appdirs==1.4.3
asgiref==3.2.7
CacheControl==0.12.6
certifi
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-random-queryset==0.1.3
django-summernote==0.8.11.6
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow
progress==1.5
psycopg2
psycopg2-binary
pyparsing==2.4.6
pytoml==0.1.21
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
